Split and refactor IDA components
Extract the IDA functionality into its own module.

Refactor to use inheritance instead of conditional class definitions.

Should allow for easier use without IDA and interoperability with other
tools.
