+++ diaphora_import.py
@@ -19,9 +19,9 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 """
 
-import diaphora
-reload(diaphora)
-from diaphora import import_definitions
+import diaphora_ida
+reload(diaphora_ida)
+from diaphora_ida import import_definitions
 
 from idaapi import IDA_SDK_VERSION
